# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AssociateProfileRequest:
    properties:
      anonProfileId:
        type: string
      realProfileId:
        type: string
    required:
    - anonProfileId
    - realProfileId
    type: object
...
  dto.FeeCalculationMethodInfo:
    properties:
      description:
        type: string
      display:
        type: string
      name:
        type: string
    type: object
  dto.FeeParticipantResponse:
    properties:
      profile:
        $ref: '#/definitions/dto.SimpleProfile'
      shareAmount:
        type: number
    type: object
...
  dto.Limits:
    properties:
      uploads:
        $ref: '#/definitions/dto.UploadLimits'
    type: object
...
  dto.NewOtherFeeRequest:
    properties:
      amount:
        type: number
      calculationMethod:
        $ref: '#/definitions/expenses.FeeCalculationMethod'
      name:
        minLength: 3
        type: string
    required:
    - amount
    - calculationMethod
    - name
    type: object
...
  dto.OtherFeeResponse:
    properties:
      amount:
        type: number
      calculationMethod:
        type: string
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      participants:
        items:
          $ref: '#/definitions/dto.FeeParticipantResponse'
        type: array
      updatedAt:
        type: string
    type: object
...
  dto.ProfileResponse:
    properties:
      associatedAnonProfileIds:
        items:
          type: string
        type: array
      avatar:
        type: string
      createdAt:
        type: string
      currentSubscription:
        $ref: '#/definitions/dto.SubscriptionResponse'
      email:
        type: string
      homeCurrency:
        type: string
      id:
        type: string
      isAnonymous:
        type: boolean
      name:
        type: string
      realProfileId:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
...
  dto.PushSubscriptionKeys:
    properties:
      auth:
        type: string
      p256dh:
        type: string
    required:
    - auth
    - p256dh
    type: object
  dto.PushSubscriptionRequest:
    properties:
      endpoint:
        type: string
      keys:
        $ref: '#/definitions/dto.PushSubscriptionKeys'
      userAgent:
        type: string
    required:
    - endpoint
    - keys
    type: object
  dto.PushUnsubscribeRequest:
    properties:
      endpoint:
        type: string
    required:
    - endpoint
    type: object
...
  dto.SimpleProfile:
    properties:
      avatar:
        type: string
      id:
        type: string
      isUser:
        type: boolean
      name:
        type: string
    type: object
  dto.SubscriptionResponse:
    properties:
      limits:
        $ref: '#/definitions/dto.Limits'
      plan:
        type: string
    type: object
...
  dto.UpdateOtherFeeRequest:
    properties:
      amount:
        type: number
      calculationMethod:
        $ref: '#/definitions/expenses.FeeCalculationMethod'
      name:
        minLength: 3
        type: string
    required:
    - amount
    - calculationMethod
    - name
    type: object
  dto.UpdateProfileRequest:
    properties:
      homeCurrency:
        type: string
      name:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - homeCurrency
    - name
    type: object
  dto.UploadLimit:
    properties:
      canUpload:
        type: boolean
      limit:
        type: integer
      remaining:
        type: integer
      resetAt:
        type: string
      used:
        type: integer
    type: object
  dto.UploadLimits:
    properties:
      canUpload:
        type: boolean
      daily:
        $ref: '#/definitions/dto.UploadLimit'
      monthly:
        $ref: '#/definitions/dto.UploadLimit'
    type: object
  expenses.FeeCalculationMethod:
    enum:
    - EQUAL_SPLIT
    - ITEMIZED_SPLIT
    type: string
    x-enum-varnames:
    - EqualSplitFee
    - ItemizedSplitFee
...
  monetization.MidtransNotificationPayload:
    properties:
      gross_amount:
        type: string
      order_id:
        type: string
      signature_key:
        type: string
      status_code:
        type: string
      status_message:
        type: string
    required:
    - gross_amount
    - order_id
    - signature_key
    - status_code
    type: object
  monetization.PaymentResponse:
    properties:
      amount:
        type: number
      createdAt:
        type: string
      currency:
        type: string
      endsAt:
        type: string
      expiredAt:
        type: string
      failureReason:
        type: string
      gateway:
        type: string
      gatewayEventId:
        type: string
      gatewaySubscriptionId:
        type: string
      gatewayTransactionId:
        type: string
      id:
        type: string
      paidAt:
        type: string
      startsAt:
        type: string
      status:
        type: string
      subscriptionId:
        type: string
      updatedAt:
        type: string
    type: object
...
  response.JSONResponse-array_dto_FeeCalculationMethodInfo:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.FeeCalculationMethodInfo'
        type: array
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
...
  response.JSONResponse-array_dto_ProfileResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ProfileResponse'
        type: array
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
...
  response.JSONResponse-dto_OtherFeeResponse:
    properties:
      data:
        $ref: '#/definitions/dto.OtherFeeResponse'
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
...
  response.JSONResponse-dto_ProfileResponse:
    properties:
      data:
        $ref: '#/definitions/dto.ProfileResponse'
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
...
  response.JSONResponse-monetization_PaymentResponse:
    properties:
      data:
        $ref: '#/definitions/monetization.PaymentResponse'
      errors:
        items: {}
        type: array
      pagination:
        $ref: '#/definitions/response.Pagination'
    type: object
...
  response.Pagination:
    properties:
      currentPage:
        type: integer
      hasNextPage:
        type: boolean
      hasPrevPage:
        type: boolean
      totalData:
        type: integer
      totalPages:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Cashback backend API
  title: Cashback API
  version: "1.0"
paths:
...
  /group-expenses/{groupExpenseId}/fees:
    post:
      consumes:
      - application/json
      parameters:
      - description: Group expense ID
        in: path
        name: groupExpenseId
        required: true
        type: string
      - description: New fee payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.NewOtherFeeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.JSONResponse-dto_OtherFeeResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Add a fee to a group expense
      tags:
      - other-fees
  /group-expenses/{groupExpenseId}/fees/{otherFeeId}:
    delete:
      parameters:
      - description: Group expense ID
        in: path
        name: groupExpenseId
        required: true
        type: string
      - description: Fee ID
        in: path
        name: otherFeeId
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove a fee from a group expense
      tags:
      - other-fees
    put:
      consumes:
      - application/json
      parameters:
      - description: Group expense ID
        in: path
        name: groupExpenseId
        required: true
        type: string
      - description: Fee ID
        in: path
        name: otherFeeId
        required: true
        type: string
      - description: Update fee payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOtherFeeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-dto_OtherFeeResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update a fee on a group expense
      tags:
      - other-fees
...
  /group-expenses/fee-calculation-methods:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-array_dto_FeeCalculationMethodInfo'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get available fee calculation methods
      tags:
      - other-fees
...
  /payments/midtrans/notifications:
    post:
      consumes:
      - application/json
      parameters:
      - description: Midtrans notification payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/monetization.MidtransNotificationPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Handle Midtrans payment notification
      tags:
      - payments
...
  /profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-dto_ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user's profile
      tags:
      - profile
    patch:
      consumes:
      - application/json
      parameters:
      - description: Update profile payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-dto_ProfileResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update current user's profile
      tags:
      - profile
  /profile/associate:
    post:
      consumes:
      - application/json
      parameters:
      - description: Associate payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AssociateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Associate anonymous profile with real profile
      tags:
      - profile
...
  /profiles:
    get:
      parameters:
      - description: Search query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.JSONResponse-array_dto_ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Search profiles
      tags:
      - profile
...
  /push/subscribe:
    post:
      consumes:
      - application/json
      parameters:
      - description: Push subscription payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.PushSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Subscribe to push notifications
      tags:
      - push
  /push/unsubscribe:
    post:
      consumes:
      - application/json
      parameters:
      - description: Push unsubscribe payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.PushUnsubscribeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Unsubscribe from push notifications
      tags:
      - push
  /subscriptions/{subscriptionId}:
    post:
      parameters:
      - description: Subscription ID
        in: path
        name: subscriptionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.JSONResponse-monetization_PaymentResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Make a payment for a subscription
      tags:
      - payments
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"